Moved latest_only operator to standard provider along with all references
