+++ airflow/example_dags/example_latest_only.py
@@ -23,7 +23,7 @@ import datetime
 
 from airflow.models.dag import DAG
 from airflow.operators.empty import EmptyOperator
-from airflow.operators.latest_only import LatestOnlyOperator
+from airflow.providers.standard.operators.latest_only import LatestOnlyOperator
 
 with DAG(
     dag_id="latest_only",
