+++ docs/apache-airflow/operators-and-hooks-ref.rst
@@ -62,7 +62,7 @@ For details see: :doc:`apache-airflow-providers:operators-and-hooks-ref/index`.
    * - :mod:`airflow.operators.generic_transfer`
      -
 
-   * - :mod:`airflow.operators.latest_only`
+   * - :mod:`airflow.providers.standard.operators.latest_only`
      -
 
    * - :mod:`airflow.providers.standard.operators.trigger_dagrun`
